// src/app/modules/operations/model/operations.model.ts
export class deviceConfig {
    dName : string="";
    dSpec : any;
    osVersion : any;
    pName : string="";
    pType : any;
    owner :  string="";
    location: string="";
    globalId : any;
    noOfInterfaces : any;
  }

  export class Device {
    devID: string="";
    prodID: any;
    location: string="";
    pName: string="";
    osType: string="";
    osVersion: any;
    cStatus: string="";
    tName: string="";
    selInterface: any;
    intAssign : any;
    lastRefresh : any;
    refreshLogs : any;
  }